♻️ refactor: 컴포넌트 이름명 변경, 공통 타입 분리
- ProductQuantityInput -> TransformQuantityInput
- ProductQuantityInputProps를 공통 타입으로 분리 (여러 곳에서 사용할 예정)
